zephyr: use mem_slab for buffers
Zephyr's k_mem_pool does not allocate buffers that are cache line
and/or DMA transfers aligned. The manual cache line manipulations
has the potential to disrupt the k_mem_pool's own tracking
information, and thus causing exceptions in freeing the memory.
So the memory allocator is modified to use mem_slab for buffers.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>